Fix Crash When Searching For Moto Actions
When we search and tap on fingerprint gestures option using the search bar given in setting, the search suggestion crashes. 

Adding android:exported=true in each activity will ensure that the activities of Moto actions can be accessed by other apps. (In this case, Settings suggestion is the app which is trying to access the fingerprint gesture activity of Moto actions app).

Signed-off-by: shivvvvvv <shivpratapd1@gmail.com>